els-ASAP auto commit file/tag genwqe-tools-4.0.20-3.el7.src.rpm

+++ src/misc/zpipe_append.c
@@ -3,7 +3,7 @@
    Version 1.4	11 December 2005  Mark Adler */
 
 /*
- * Copyright 2016, International Business Machines
+ * Copyright 2016, 2017 International Business Machines
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -100,7 +100,7 @@ static int def(FILE *source, FILE *dest, int window_bits, int _flush,
 		return Z_ERRNO;
 
 	out = malloc(CHUNK_o);
-	if (in == NULL) {
+	if (out == NULL) {
 		free(in);
 		return Z_ERRNO;
 	}
@@ -113,8 +113,11 @@ static int def(FILE *source, FILE *dest, int window_bits, int _flush,
 
 	ret = deflateInit2(&strm, level, Z_DEFLATED, window_bits, 8,
 			   Z_DEFAULT_STRATEGY);
-	if (ret != Z_OK)
+	if (ret != Z_OK) {
+		free(in);
+		free(out);
 		return ret;
+	}
 
 	/* compress until end of file */
 	do {
@@ -210,7 +213,7 @@ static int inf(FILE *source, FILE *dest, int window_bits, int _flush,
 		return Z_ERRNO;
 
 	out = malloc(CHUNK_o);
-	if (in == NULL) {
+	if (out == NULL) {
 		free(in);
 		return Z_ERRNO;
 	}
@@ -276,7 +279,10 @@ static int inf(FILE *source, FILE *dest, int window_bits, int _flush,
 
 			switch (ret) {
 			case Z_NEED_DICT:
-				ret = Z_DATA_ERROR;	/* and fall through */
+				(void)inflateEnd(&strm);
+				free(in);
+				free(out);
+				return Z_DATA_ERROR;
 			case Z_DATA_ERROR:
 			case Z_MEM_ERROR:
 				(void)inflateEnd(&strm);
@@ -372,13 +378,14 @@ static void usage(char *prog)
 
 	fprintf(stderr, "%s usage: %s [-h] [-v]\n"
 		"    [-F, --format <ZLIB|DEFLATE|GZIP>]\n"
-		"    [-e, --excact-input]  input matches size of data\n"
+		"    [-e, --excact-input] input matches size of data\n"
 		"    [-E, --excact-output] output matches size of data\n"
 		"    [-f, --fush <Z_NO_FLUSH|Z_PARTIAL_FLUSH|Z_FULL_FLUSH>]\n"
 		"    [-i, --i_bufsize <i_bufsize>]\n"
 		"    [-o, --o_bufsize <o_bufsize>]\n"
 		"    [-p, --pattern <pattern>] pattern to generate test-data\n"
-		"    [-s, --size <data-size>]\n",
+		"    [-s, --size <data-size>]\n"
+		"    [-k, --keep] do not delete resulting files\n",
 		b, b);
 }
 
@@ -397,6 +404,7 @@ int main(int argc, char **argv)
 	int exact_input = 0, exact_output = 0;
 	size_t size = 256 * 1024;
 	int expect_z_stream_end = 0;
+	int keep = 0;
 
 	_pattern = getpid();
 
@@ -416,12 +424,13 @@ int main(int argc, char **argv)
 			{ "o_bufsize",	 required_argument, NULL, 'o' },
 			{ "size",	 required_argument, NULL, 's' },
 			{ "pattern",	 required_argument, NULL, 'p' },
+			{ "keep",	 no_argument,       NULL, 'k' },
 			{ "verbose",	 no_argument,	    NULL, 'v' },
 			{ "help",	 no_argument,	    NULL, 'h' },
 			{ 0,		 no_argument,	    NULL, 0   },
 		};
 
-		ch = getopt_long(argc, argv, "F:f:Eei:o:s:p:vh?",
+		ch = getopt_long(argc, argv, "kF:f:Eei:o:s:p:vh?",
 				 long_options, &option_index);
 		if (ch == -1)    /* all params processed ? */
 			break;
@@ -446,6 +455,9 @@ int main(int argc, char **argv)
 			break;
 		case 'E':
 			exact_output = 1;
+			break;
+		case 'k':
+			keep = 1;
 			break;
 		case 'v':
 			verbose++;
@@ -503,6 +515,8 @@ int main(int argc, char **argv)
 	rc = def(i_fp, o_fp, window_bits, flush, Z_DEFAULT_COMPRESSION,
 		 &expected_bytes, &decompressed_bytes);
 	if (rc != Z_OK) {
+		fclose(o_fp);
+		fclose(i_fp);
 		fprintf(stderr, "err: compression failed.\n");
 		zerr(rc);
 		return rc;
@@ -566,6 +580,8 @@ int main(int argc, char **argv)
 	}
 
 	if (rc != Z_OK) {
+		fclose(o_fp);
+		fclose(n_fp);
 		fprintf(stderr, "err: decompression failed.\n");
 		zerr(rc);
 		return rc;
@@ -582,9 +598,11 @@ int main(int argc, char **argv)
 		exit(EXIT_FAILURE);
 	}
 
-	unlink(i_fname);
-	unlink(n_fname);
-	unlink(o_fname);
+	if (!keep) {
+		unlink(i_fname);
+		unlink(n_fname);
+		unlink(o_fname);
+	}
 
 	exit(EXIT_SUCCESS);
 }
